Moving ADC reading to helper function file. Hopefully to be expanded to tidy up other code.
